freeimage: Unbundle dependencies, fix AAarch64 support, and other fixes
This pretty much just adapts https://src.fedoraproject.org/rpms/freeimage into nixpkgs.
All of the patches except the substream one and doxygen one is included, since the
purpose of the first one is unclear and the second one is just for docs.

It seems to me that freeimage is unfortunately partly abandonware, considering
that there is a patch for a CVE included and that the last release was in 2018.

Fixes #77653
Replaces #77655