python3Packages.selenium: purge source of bundled binaries before building
the build process already replaces the relevant .so with one from
nixpkgs, but only for the current architecture, leaving the "other"
architecture's pre-built .so being copied to the output. saves a
few k in the output size in the process.

this is an alternative to slapping a sourceProvenance binaryNativeCode
on it